Assays objects

The Assays virtual class and its methods provide a formal abstraction of the assays slot of SummarizedExperiment objects.

SimpleAssays and ShallowSimpleListAssays are concrete subclasses of Assays with the former being currently the default implementation of Assays objects. Other implementations (e.g. disk-based) could easily be added.

Note that these classes are not meant to be used directly by the end user and the material in this man page is aimed at package developers.

Details

Assays objects have a list-like semantics with elements having matrix- or array-like semantics (e.g., dim, dimnames).

The Assays API consists of:

(a) The Assays() constructor function.

(b) Lossless back and forth coercion from/to SimpleList. The coercion method from SimpleList doesn't need (and should not) validate the returned object.

(c) length, names, `names<-`, getListElement, setListElement, dim, [, `[<-`, rbind, cbind.

An Assays concrete subclass needs to implement (b) (required) plus, optionally any of the methods in (c).

IMPORTANT:

Nobody in the Assays hierarchy is allowed to inherit from SimpleList because of the conflicting semantic of [.

Methods that return a modified Assays object (a.k.a. endomorphisms), that is, [ as well as replacement methods names<-, setListElement, and [<-, must respect the copy-on-change contract. With objects that don't make use of references internally, the developer doesn't need to take any special action for that because it's automatically taken care of by R itself. However, for objects that do make use of references internally (e.g. environments, external pointers, pointer to a file on disk, etc...), the developer needs to be careful to implement endomorphisms with copy-on-change semantics. This can easily be achieved (and is what the default methods for Assays objects do) by performaing a full (deep) copy of the object before modifying it instead of trying to modify it in-place. However note that this full (deep) copy can be very expensive and is actually not necessary in order to achieve copy-on-change semantics: it's enough (and often preferrable for performance reasons) to copy only the parts of the object that need to be modified.

Assays has currently 3 implementations which are formalized by concrete subclasses SimpleAssays, ShallowSimpleListAssays, and AssaysInEnv. SimpleAssays is the default (prior to SummarizedExperiment 1.15.4, ShallowSimpleListAssays was the default). AssaysInEnv is a broken alternative to ShallowSimpleListAssays that does NOT respect the copy-on-change contract. It is only provided for illustration purposes (see source file Assays-class.R for the details).

A little more detail about ShallowSimpleListAssays: a small reference class hierarchy (not exported from the GenomicRanges name space) defines a reference class ShallowData with a single field data of type ANY, and a derived class ShallowSimpleListAssays that specializes the type of data as SimpleList, and contains=c("ShallowData", "Assays"). The assays slot of a SummarizedExperiment object contains an instance of ShallowSimpleListAssays.

Author

Martin Morgan and Hervé Pagès

See also

Examples

## ---------------------------------------------------------------------
## DIRECT MANIPULATION OF Assays OBJECTS
## ---------------------------------------------------------------------
m1 <- matrix(runif(24), ncol=3)
m2 <- matrix(runif(24), ncol=3)
a <- Assays(SimpleList(m1, m2))
a
#> An object of class "SimpleAssays"
#> Slot "data":
#> List of length 2
#> 

as(a, "SimpleList")
#> List of length 2

length(a)
#> [1] 2
getListElement(a, 2)
#>            [,1]      [,2]       [,3]
#> [1,] 0.98053967 0.3008308 0.68016292
#> [2,] 0.74152153 0.6364656 0.49884561
#> [3,] 0.05144628 0.4790245 0.64167935
#> [4,] 0.53021246 0.4321713 0.66028435
#> [5,] 0.69582388 0.7064338 0.09602416
#> [6,] 0.68855600 0.9485766 0.76560016
#> [7,] 0.03123033 0.1803388 0.76967480
#> [8,] 0.22556253 0.2168999 0.99071231
dim(a)
#> [1] 8 3

b <- a[-4, 2]
b
#> An object of class "SimpleAssays"
#> Slot "data":
#> List of length 2
#> 
length(b)
#> [1] 2
getListElement(b, 2)
#>           [,1]
#> [1,] 0.3008308
#> [2,] 0.6364656
#> [3,] 0.4790245
#> [4,] 0.7064338
#> [5,] 0.9485766
#> [6,] 0.1803388
#> [7,] 0.2168999
dim(b)
#> [1] 7 1

names(a)
#> NULL
names(a) <- c("a1", "a2")
names(a)
#> [1] "a1" "a2"
getListElement(a, "a2")
#>            [,1]      [,2]       [,3]
#> [1,] 0.98053967 0.3008308 0.68016292
#> [2,] 0.74152153 0.6364656 0.49884561
#> [3,] 0.05144628 0.4790245 0.64167935
#> [4,] 0.53021246 0.4321713 0.66028435
#> [5,] 0.69582388 0.7064338 0.09602416
#> [6,] 0.68855600 0.9485766 0.76560016
#> [7,] 0.03123033 0.1803388 0.76967480
#> [8,] 0.22556253 0.2168999 0.99071231

rbind(a, a)
#> An object of class "SimpleAssays"
#> Slot "data":
#> List of length 2
#> names(2): a1 a2
#> 
cbind(a, a)
#> An object of class "SimpleAssays"
#> Slot "data":
#> List of length 2
#> names(2): a1 a2
#> 

## ---------------------------------------------------------------------
## COPY-ON-CHANGE CONTRACT
## ---------------------------------------------------------------------

## ShallowSimpleListAssays objects have copy-on-change semantics but not
## AssaysInEnv objects. For example:
ssla <- as(SimpleList(m1, m2), "ShallowSimpleListAssays")
aie <- as(SimpleList(m1, m2), "AssaysInEnv")

## No names on 'ssla' and 'aie':
names(ssla)
#> NULL
names(aie)
#> NULL

ssla2 <- ssla
aie2 <- aie
names(ssla2) <- names(aie2) <- c("A1", "A2")

names(ssla)  # still NULL (as expected)
#> NULL

names(aie)   # changed! (because the names<-,AssaysInEnv method is not
#> [1] "A1" "A2"
             # implemented in a way that respects the copy-on-change
             # contract)