RangedSummarizedExperiment objects

The RangedSummarizedExperiment class is a matrix-like container where rows represent ranges of interest (as a GRanges or GRangesList object) and columns represent samples (with sample data summarized as a DataFrame). A RangedSummarizedExperiment contains one or more assays, each represented by a matrix-like object of numeric or other mode.

RangedSummarizedExperiment is a subclass of SummarizedExperiment and, as such, all the methods documented in class?SummarizedExperiment also work on a RangedSummarizedExperiment object. The methods documented below are additional methods that are specific to RangedSummarizedExperiment objects.

## Constructor

SummarizedExperiment(assays=SimpleList(),
                     rowData=NULL, rowRanges=GRangesList(),
                     colData=DataFrame(),
                     metadata=list(),
                     checkDimnames=TRUE)

## Accessors

rowRanges(x, ...)
rowRanges(x, ...) <- value

## Subsetting

# S4 method for RangedSummarizedExperiment
subset(x, subset, select, ...)

## rowRanges access
## see 'GRanges compatibility', below

Arguments

assays
A list or SimpleList of matrix-like elements, or a matrix-like object (e.g. an ordinary matrix, a data frame, a DataFrame object from the S4Vectors package, a sparseMatrix derivative from the Matrix package, a DelayedMatrix object from the DelayedArray package, etc...). All elements of the list must have the same dimensions, and dimension names (if present) must be consistent across elements and with the row names of rowRanges and colData.

rowData
A DataFrame object describing the rows. Row names, if present, become the row names of the SummarizedExperiment object. The number of rows of the DataFrame must equal the number of rows of the matrices in assays.

rowRanges
A GRanges or GRangesList object describing the ranges of interest. Names, if present, become the row names of the SummarizedExperiment object. The length of the GRanges or GRangesList must equal the number of rows of the matrices in assays. If rowRanges is missing, a SummarizedExperiment instance is returned.

colData
An optional DataFrame describing the samples. Row names, if present, become the column names of the RangedSummarizedExperiment.

metadata
An optional list of arbitrary content describing the overall experiment.

checkDimnames
By default the rownames and colnames of the supplied assay(s) are checked for consistency with those of the SummarizedExperiment object (or derivative) to construct. More precisely, the rownames and colnames of each assay must be NULL or identical to those of the object. Use checkDimnames=FALSE to skip this check.

x
A RangedSummarizedExperiment object. The rowRanges setter will also accept a SummarizedExperiment object and will first coerce it to RangedSummarizedExperiment before it sets value on it.

...
Further arguments to be passed to or from other methods.

value
A GRanges or GRangesList object.

subset
An expression which, when evaluated in the context of rowRanges(x), is a logical vector indicating elements or rows to keep: missing values are taken as false.

select
An expression which, when evaluated in the context of colData(x), is a logical vector indicating elements or rows to keep: missing values are taken as false.

Details

The rows of a RangedSummarizedExperiment object represent ranges (in genomic coordinates) of interest. The ranges of interest are described by a GRanges or a GRangesList object, accessible using the rowRanges function, described below. The GRanges and GRangesList classes contains sequence (e.g., chromosome) name, genomic coordinates, and strand information. Each range can be annotated with additional data; this data might be used to describe the range or to summarize results (e.g., statistics of differential abundance) relevant to the range. Rows may or may not have row names; they often will not.

Constructor

RangedSummarizedExperiment instances are constructed using the SummarizedExperiment() function with arguments outlined above.

Accessors

In the following code snippets, x is a RangedSummarizedExperiment object.

rowRanges(x), rowRanges(x) <- value:

Get or set the row data. value is a GenomicRanges object. Row names of value must be NULL or consistent with the existing row names of x.

GRanges compatibility (rowRanges access)

Many GRanges and GRangesList operations are supported on RangedSummarizedExperiment objects, using rowRanges.

Supported operations include: pcompare, duplicated, end, end<-, granges, is.unsorted, match, mcols, mcols<-, order, ranges, ranges<-, rank, seqinfo, seqinfo<-, seqnames, sort, start, start<-, strand, strand<-, width, width<-.

See also ?shift, ?isDisjoint, ?coverage, ?findOverlaps, and ?nearest for more GRanges compatibility methods.

Not all GRanges operations are supported, because they do not make sense for RangedSummarizedExperiment objects (e.g., length, name, as.data.frame, c, splitAsList), involve non-trivial combination or splitting of rows (e.g., disjoin, gaps, reduce, unique), or have not yet been implemented (Ops, map, window, window<-).

Subsetting

In the code snippets below, x is a RangedSummarizedExperiment object.

subset(x, subset, select):

Create a subset of x using an expression subset referring to columns of rowRanges(x) (including ‘seqnames’, ‘start’, ‘end’, ‘width’, ‘strand’, and names(rowData(x))) and / or select referring to column names of colData(x).

Extension

RangedSummarizedExperiment is implemented as an S4 class, and can be extended in the usual way, using contains="RangedSummarizedExperiment" in the new class definition.

Author

Martin Morgan, mtmorgan@fhcrc.org

See also

Examples

nrows <- 200; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
                     IRanges(floor(runif(200, 1e5, 1e6)), width=100),
                     strand=sample(c("+", "-"), 200, TRUE),
                     feature_id=sprintf("ID%03d", 1:200))
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            rowRanges=rowRanges, colData=colData)
rse
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
dim(rse)
#> [1] 200   6
dimnames(rse)
#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F"
#> 
assayNames(rse)
#> [1] "counts"
head(assay(rse))
#>             A        B        C         D          E        F
#> [1,] 9705.239 5428.829 1591.189 9832.6854   17.46324 7087.612
#> [2,] 3892.438 2393.642 2938.214 1091.3788 4180.24239 3752.030
#> [3,] 4612.403 5089.427 8912.957  700.3852 7639.57967 2548.483
#> [4,] 3153.102 4173.226 5881.198 7776.6774 7944.98219 9964.133
#> [5,] 1747.584 7269.762 6332.682 6583.5640 2426.62762 5229.431
#> [6,] 5316.204 6377.218 2595.835 2592.7682 7591.95037 7158.378
assays(rse) <- endoapply(assays(rse), asinh)
head(assay(rse))
#>             A        B        C        D        E        F
#> [1,] 9.873568 9.292626 8.065384 9.886615 3.554064 9.559251
#> [2,] 8.959938 8.473718 8.678704 7.688345 9.031272 8.923199
#> [3,] 9.129652 9.228068 9.788409 7.244778 9.634245 8.536401
#> [4,] 8.749289 9.029592 9.372663 9.652032 9.673443 9.899894
#> [5,] 8.159137 9.584626 9.446626 9.485479 8.487405 9.255205
#> [6,] 9.271662 9.453634 8.554811 8.553629 9.627991 9.569186

rowRanges(rse)
#> GRanges object with 200 ranges and 1 metadata column:
#>         seqnames        ranges strand |  feature_id
#>            <Rle>     <IRanges>  <Rle> | <character>
#>     [1]     chr1 274328-274427      + |       ID001
#>     [2]     chr1 382760-382859      - |       ID002
#>     [3]     chr1 223779-223878      + |       ID003
#>     [4]     chr1 863022-863121      - |       ID004
#>     [5]     chr1 536699-536798      - |       ID005
#>     ...      ...           ...    ... .         ...
#>   [196]     chr2 755844-755943      - |       ID196
#>   [197]     chr2 704900-704999      + |       ID197
#>   [198]     chr2 837820-837919      - |       ID198
#>   [199]     chr2 923159-923258      + |       ID199
#>   [200]     chr2 720973-721072      + |       ID200
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
rowData(rse)  # same as 'mcols(rowRanges(rse))'
#> DataFrame with 200 rows and 1 column
#>      feature_id
#>     <character>
#> 1         ID001
#> 2         ID002
#> 3         ID003
#> 4         ID004
#> 5         ID005
#> ...         ...
#> 196       ID196
#> 197       ID197
#> 198       ID198
#> 199       ID199
#> 200       ID200
colData(rse)
#> DataFrame with 6 rows and 1 column
#>     Treatment
#>   <character>
#> A        ChIP
#> B       Input
#> C        ChIP
#> D       Input
#> E        ChIP
#> F       Input

rse[ , rse$Treatment == "ChIP"]
#> class: RangedSummarizedExperiment 
#> dim: 200 3 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(3): A C E
#> colData names(1): Treatment

## cbind() combines objects with the same ranges but different samples:
rse1 <- rse
rse2 <- rse1[ , 1:3]
colnames(rse2) <- letters[1:ncol(rse2)] 
cmb1 <- cbind(rse1, rse2)
dim(cmb1)
#> [1] 200   9
dimnames(cmb1)
#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F" "a" "b" "c"
#> 

## rbind() combines objects with the same samples but different ranges:
rse1 <- rse
rse2 <- rse1[1:50, ]
rownames(rse2) <- letters[1:nrow(rse2)] 
cmb2 <- rbind(rse1, rse2)
dim(cmb2)
#> [1] 250   6
dimnames(cmb2)
#> [[1]]
#>   [1] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [19] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [37] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [55] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [73] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [91] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [109] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [127] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [145] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [163] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [181] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [199] ""  ""  "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p"
#> [217] "q" "r" "s" "t" "u" "v" "w" "x" "y" "z" NA  NA  NA  NA  NA  NA  NA  NA 
#> [235] NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA 
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F"
#> 

## Coercion to/from SummarizedExperiment:
se0 <- as(rse, "SummarizedExperiment")
se0
#> class: SummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment

as(se0, "RangedSummarizedExperiment")
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment

## Setting rowRanges on a SummarizedExperiment object turns it into a
## RangedSummarizedExperiment object:
se <- se0
rowRanges(se) <- rowRanges
se  # RangedSummarizedExperiment
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment

## Sanity checks:
stopifnot(identical(assays(se0), assays(rse)))
stopifnot(identical(dim(se0), dim(rse)))
stopifnot(identical(dimnames(se0), dimnames(rse)))
stopifnot(identical(rowData(se0), rowData(rse)))
stopifnot(identical(colData(se0), colData(rse)))