Inter range transformations of a RangedSummarizedExperiment object

This man page documents the inter range transformations that are supported on RangedSummarizedExperiment objects.

# S4 method for RangedSummarizedExperiment
isDisjoint(x, ignore.strand=FALSE)

# S4 method for RangedSummarizedExperiment
disjointBins(x, ignore.strand=FALSE)

Arguments

x
A RangedSummarizedExperiment object.

ignore.strand
See ?isDisjoint in the GenomicRanges package.

Details

These transformations operate on the rowRanges component of the RangedSummarizedExperiment object, which can be a GenomicRanges or GRangesList object.

More precisely, any of the above functions performs the following transformation on RangedSummarizedExperiment object x:

    f(rowRanges(x), ...)

where f is the name of the function and ... any additional arguments passed to it.

See ?isDisjoint in the GenomicRanges package for the details of how these transformations operate on a GenomicRanges or GRangesList object.

Value

See ?isDisjoint in the GenomicRanges package.

See also

Examples

nrows <- 20; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(5, 15)),
                     IRanges(sample(1000L, 20), width=100),
                     strand=Rle(c("+", "-"), c(12, 8)))
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
rse0 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                             rowRanges=rowRanges, colData=colData)
rse1 <- shift(rse0, 99*start(rse0))

isDisjoint(rse0)  # FALSE
#> [1] FALSE
isDisjoint(rse1)  # TRUE
#> [1] TRUE

bins0 <- disjointBins(rse0)
bins0
#>  [1] 1 1 1 1 1 1 3 1 1 1 1 2 1 3 1 1 2 1 1 2
stopifnot(identical(bins0, disjointBins(rowRanges(rse0))))

bins1 <- disjointBins(rse1)
bins1
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
stopifnot(all(bins1 == bins1[1]))