Make a RangedSummarizedExperiment object from an ExpressionSet and vice-versa

Coercion between RangedSummarizedExperiment and ExpressionSet is supported in both directions.

For going from ExpressionSet to RangedSummarizedExperiment, the makeSummarizedExperimentFromExpressionSet function is also provided to let the user control how to map features to ranges.

makeSummarizedExperimentFromExpressionSet(from,
                                          mapFun=naiveRangeMapper,
                                          ...)

## range mapping functions
naiveRangeMapper(from)
probeRangeMapper(from)
geneRangeMapper(txDbPackage, key = "ENTREZID")

Arguments

from
An ExpressionSet object.

mapFun
A function which takes an ExpressionSet object and returns a GRanges, or GRangesList object which corresponds to the genomic ranges used in the ExpressionSet. The rownames of the returned GRanges are used to match the featureNames of the ExpressionSet.

The naiveRangeMapper function is used by default.

...
Additional arguments passed to mapFun.

txDbPackage
A character string with the Transcript Database to use for the mapping.

key
A character string with the Gene key to use for the mapping.

Value

makeSummarizedExperimentFromExpressionSet takes an ExpressionSet object as input and a range mapping function that maps the features to ranges. It then returns a RangedSummarizedExperiment object that corresponds to the input.

The range mapping functions return a GRanges object, with the rownames corresponding to the featureNames of the ExpressionSet object.

Author

Jim Hester, james.f.hester@gmail.com

See also

Examples

## ---------------------------------------------------------------------
## GOING FROM ExpressionSet TO SummarizedExperiment
## ---------------------------------------------------------------------

data(sample.ExpressionSet, package="Biobase")

# naive coercion
makeSummarizedExperimentFromExpressionSet(sample.ExpressionSet)
#> class: RangedSummarizedExperiment 
#> dim: 500 26 
#> metadata(3): experimentData annotation protocolData
#> assays(2): exprs se.exprs
#> rownames(500): AFFX-MurIL2_at AFFX-MurIL10_at ... 31738_at 31739_at
#> rowData names(0):
#> colnames(26): A B ... Y Z
#> colData names(3): sex type score
as(sample.ExpressionSet, "RangedSummarizedExperiment")
#> class: RangedSummarizedExperiment 
#> dim: 500 26 
#> metadata(3): experimentData annotation protocolData
#> assays(2): exprs se.exprs
#> rownames(500): AFFX-MurIL2_at AFFX-MurIL10_at ... 31738_at 31739_at
#> rowData names(0):
#> colnames(26): A B ... Y Z
#> colData names(3): sex type score
as(sample.ExpressionSet, "SummarizedExperiment")
#> class: SummarizedExperiment 
#> dim: 500 26 
#> metadata(3): experimentData annotation protocolData
#> assays(2): exprs se.exprs
#> rownames(500): AFFX-MurIL2_at AFFX-MurIL10_at ... 31738_at 31739_at
#> rowData names(0):
#> colnames(26): A B ... Y Z
#> colData names(3): sex type score

# using probe range mapper
makeSummarizedExperimentFromExpressionSet(sample.ExpressionSet, probeRangeMapper)
#> 
#> 
#> Warning: Accessing gene location information via 'CHR','CHRLOC','CHRLOCEND' is
#>   deprecated. Please use a range based accessor like genes(), or select()
#>   with columns values like TXCHROM and TXSTART on a TxDb or OrganismDb
#>   object instead.
#> 'select()' returned 1:many mapping between keys and columns
#> Warning: 148 probes could not be mapped.
#> class: RangedSummarizedExperiment 
#> dim: 352 26 
#> metadata(3): experimentData annotation protocolData
#> assays(2): exprs se.exprs
#> rownames(352): AFFX-HUMISGF3A/M97935_5_at AFFX-HUMISGF3A/M97935_MA_at
#>   ... 31736_at 31737_at
#> rowData names(0):
#> colnames(26): A B ... Y Z
#> colData names(3): sex type score

# using the gene range mapper
se <- makeSummarizedExperimentFromExpressionSet(
    sample.ExpressionSet,
    geneRangeMapper("TxDb.Hsapiens.UCSC.hg19.knownGene")
)
#> 'select()' returned 1:many mapping between keys and columns
#> Registered S3 methods overwritten by 'SummarizedExperiment':
#>   method                                  from                   
#>   .DollarNames.SummarizedExperiment       SummarizedExperimentGHA
#>   .DollarNames.RangedSummarizedExperiment SummarizedExperimentGHA
#>   403 genes were dropped because they have exons located on both strands
#>   of the same reference sequence or on more than one reference sequence,
#>   so cannot be represented by a single genomic range.
#>   Use 'single.strand.genes.only=FALSE' to get all the genes in a
#>   GRangesList object, or use suppressMessages() to suppress this message.
#> Warning: 171 probes could not be mapped.
se
#> class: RangedSummarizedExperiment 
#> dim: 329 26 
#> metadata(3): experimentData annotation protocolData
#> assays(2): exprs se.exprs
#> rownames(329): AFFX-HUMISGF3A/M97935_5_at AFFX-HUMISGF3A/M97935_5_at
#>   ... 31736_at 31737_at
#> rowData names(1): gene_id
#> colnames(26): A B ... Y Z
#> colData names(3): sex type score
rowData(se)  # duplicate row names
#> DataFrame with 329 rows and 1 column
#>                                gene_id
#>                            <character>
#> AFFX-HUMISGF3A/M97935_5_at        6772
#> AFFX-HUMISGF3A/M97935_5_at        6772
#> AFFX-HUMISGF3A/M97935_5_at        6772
#> AFFX-HUMISGF3A/M97935_5_at        6772
#> AFFX-HUMGAPDH/M33197_5_at         2597
#> ...                                ...
#> 31733_at                          5024
#> 31734_at                          3227
#> 31735_at                         80316
#> 31736_at                          6801
#> 31737_at                            58

## ---------------------------------------------------------------------
## GOING FROM SummarizedExperiment TO ExpressionSet
## ---------------------------------------------------------------------

example(RangedSummarizedExperiment)  # to create 'rse'
#> 
#> RngdSE> nrows <- 200; ncols <- 6
#> 
#> RngdSE> counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
#> 
#> RngdSE> rowRanges <- GRanges(rep(c("chr1", "chr2"), c(50, 150)),
#> RngdSE+                      IRanges(floor(runif(200, 1e5, 1e6)), width=100),
#> RngdSE+                      strand=sample(c("+", "-"), 200, TRUE),
#> RngdSE+                      feature_id=sprintf("ID%03d", 1:200))
#> 
#> RngdSE> colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
#> RngdSE+                      row.names=LETTERS[1:6])
#> 
#> RngdSE> rse <- SummarizedExperiment(assays=SimpleList(counts=counts),
#> RngdSE+                             rowRanges=rowRanges, colData=colData)
#> 
#> RngdSE> rse
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
#> 
#> RngdSE> dim(rse)
#> [1] 200   6
#> 
#> RngdSE> dimnames(rse)
#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F"
#> 
#> 
#> RngdSE> assayNames(rse)
#> [1] "counts"
#> 
#> RngdSE> head(assay(rse))
#>             A        B          C        D        E        F
#> [1,] 4722.245 6566.958 9152.82308 1949.327 3179.950 2728.438
#> [2,] 2826.128 4538.774 4992.65080 2735.593 8332.091 1554.358
#> [3,] 8319.229 9138.503 2663.41726 9332.805 6347.146 8032.617
#> [4,] 5298.989 8171.587 9627.21522 3752.260 3154.935 3069.180
#> [5,] 6627.937 8182.028   30.76941 4008.576 1366.718 2004.853
#> [6,] 5236.918 4934.338 8488.34019 6210.713 1490.145 5358.842
#> 
#> RngdSE> assays(rse) <- endoapply(assays(rse), asinh)
#> 
#> RngdSE> head(assay(rse))
#>             A        B        C        D        E        F
#> [1,] 9.153187 9.482953 9.814965 8.268387 8.757768 8.604632
#> [2,] 8.639810 9.113559 9.208869 8.607251 9.721017 8.041965
#> [3,] 9.719472 9.813399 8.580512 9.834438 9.448908 9.684413
#> [4,] 9.268419 9.701566 9.865496 8.923261 8.749870 8.722313
#> [5,] 9.492196 9.702842 4.119932 8.989339 7.913315 8.296473
#> [6,] 9.256636 9.197121 9.739596 9.427178 7.999776 9.279650
#> 
#> RngdSE> rowRanges(rse)
#> GRanges object with 200 ranges and 1 metadata column:
#>         seqnames        ranges strand |  feature_id
#>            <Rle>     <IRanges>  <Rle> | <character>
#>     [1]     chr1 913405-913504      - |       ID001
#>     [2]     chr1 397194-397293      - |       ID002
#>     [3]     chr1 466379-466478      - |       ID003
#>     [4]     chr1 936071-936170      + |       ID004
#>     [5]     chr1 736514-736613      + |       ID005
#>     ...      ...           ...    ... .         ...
#>   [196]     chr2 279383-279482      + |       ID196
#>   [197]     chr2 765902-766001      - |       ID197
#>   [198]     chr2 129822-129921      - |       ID198
#>   [199]     chr2 921802-921901      - |       ID199
#>   [200]     chr2 866617-866716      - |       ID200
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> RngdSE> rowData(rse)  # same as 'mcols(rowRanges(rse))'
#> DataFrame with 200 rows and 1 column
#>      feature_id
#>     <character>
#> 1         ID001
#> 2         ID002
#> 3         ID003
#> 4         ID004
#> 5         ID005
#> ...         ...
#> 196       ID196
#> 197       ID197
#> 198       ID198
#> 199       ID199
#> 200       ID200
#> 
#> RngdSE> colData(rse)
#> DataFrame with 6 rows and 1 column
#>     Treatment
#>   <character>
#> A        ChIP
#> B       Input
#> C        ChIP
#> D       Input
#> E        ChIP
#> F       Input
#> 
#> RngdSE> rse[ , rse$Treatment == "ChIP"]
#> class: RangedSummarizedExperiment 
#> dim: 200 3 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(3): A C E
#> colData names(1): Treatment
#> 
#> RngdSE> ## cbind() combines objects with the same ranges but different samples:
#> RngdSE> rse1 <- rse
#> 
#> RngdSE> rse2 <- rse1[ , 1:3]
#> 
#> RngdSE> colnames(rse2) <- letters[1:ncol(rse2)] 
#> 
#> RngdSE> cmb1 <- cbind(rse1, rse2)
#> 
#> RngdSE> dim(cmb1)
#> [1] 200   9
#> 
#> RngdSE> dimnames(cmb1)
#> [[1]]
#> NULL
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F" "a" "b" "c"
#> 
#> 
#> RngdSE> ## rbind() combines objects with the same samples but different ranges:
#> RngdSE> rse1 <- rse
#> 
#> RngdSE> rse2 <- rse1[1:50, ]
#> 
#> RngdSE> rownames(rse2) <- letters[1:nrow(rse2)] 
#> 
#> RngdSE> cmb2 <- rbind(rse1, rse2)
#> 
#> RngdSE> dim(cmb2)
#> [1] 250   6
#> 
#> RngdSE> dimnames(cmb2)
#> [[1]]
#>   [1] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [19] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [37] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [55] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [73] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#>  [91] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [109] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [127] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [145] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [163] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [181] ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  ""  "" 
#> [199] ""  ""  "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "n" "o" "p"
#> [217] "q" "r" "s" "t" "u" "v" "w" "x" "y" "z" NA  NA  NA  NA  NA  NA  NA  NA 
#> [235] NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA 
#> 
#> [[2]]
#> [1] "A" "B" "C" "D" "E" "F"
#> 
#> 
#> RngdSE> ## Coercion to/from SummarizedExperiment:
#> RngdSE> se0 <- as(rse, "SummarizedExperiment")
#> 
#> RngdSE> se0
#> class: SummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
#> 
#> RngdSE> as(se0, "RangedSummarizedExperiment")
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
#> 
#> RngdSE> ## Setting rowRanges on a SummarizedExperiment object turns it into a
#> RngdSE> ## RangedSummarizedExperiment object:
#> RngdSE> se <- se0
#> 
#> RngdSE> rowRanges(se) <- rowRanges
#> 
#> RngdSE> se  # RangedSummarizedExperiment
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
#> 
#> RngdSE> ## Sanity checks:
#> RngdSE> stopifnot(identical(assays(se0), assays(rse)))
#> 
#> RngdSE> stopifnot(identical(dim(se0), dim(rse)))
#> 
#> RngdSE> stopifnot(identical(dimnames(se0), dimnames(rse)))
#> 
#> RngdSE> stopifnot(identical(rowData(se0), rowData(rse)))
#> 
#> RngdSE> stopifnot(identical(colData(se0), colData(rse)))
rse
#> class: RangedSummarizedExperiment 
#> dim: 200 6 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(1): feature_id
#> colnames(6): A B ... E F
#> colData names(1): Treatment
as(rse, "ExpressionSet")
#> Warning: No assay named ‘exprs’ found, renaming counts to ‘exprs’.
#> ExpressionSet (storageMode: lockedEnvironment)
#> assayData: 200 features, 6 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: A B ... F (6 total)
#>   varLabels: Treatment
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: 1 2 ... 200 (200 total)
#>   fvarLabels: seqnames start ... feature_id (6 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation: