Make a SummarizedExperiment from a '.loom' hdf5 file

makeSummarizedExperimentFromLoom represents a '.loom' file as a SummarizedExperiment. The '/matrix' and '/layers' are represented as HDF5Array objects; row and column attributes are parsed to DataFrame. Optionally, row or column attributes can be specified as row and and column names.

makeSummarizedExperimentFromLoom(file,
                                 rownames_attr = NULL,
                                 colnames_attr = NULL)

Arguments

file
The path (as a single character string) to the HDF5 file where the dataset is located.

rownames_attr
The name of the row attribute to be used as row names.

colnames_attr
The name of the column attribute to be used as column names.

Value

A SummarizedExperiment object with row and column data and one or more assays.

Author

Martin Morgan

See also

Examples

## ---------------------------------------------------------------------
## BASIC EXAMPLE
## ---------------------------------------------------------------------

file <- system.file(
    package="SummarizedExperiment", "extdata", "example.loom"
)
se <- makeSummarizedExperimentFromLoom(file)
se
#> class: SummarizedExperiment 
#> dim: 20 6 
#> metadata(1): title
#> assays(3): matrix a b
#> rownames: NULL
#> rowData names(2): id score
#> colnames: NULL
#> colData names(2): group id
assay(se)
#> <20 x 6> matrix of class DelayedMatrix and type "double":
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,] -0.60276831 -1.30311075  0.61454830  0.05457548 -0.65739079 -1.09495208
#>  [2,]  0.58371117 -0.04277452  0.18922219  1.11331417 -0.53579363  0.66678297
#>  [3,]  1.67964076  1.12830991  0.19522145  0.31074607 -0.48807881  0.65210311
#>  [4,] -0.16591875 -1.32472050  0.41129405 -0.63319603  1.73252508  0.69186273
#>  [5,]  0.73705399  1.21444274  0.20758942  0.42888040  1.12116331 -0.65071886
#>   ...           .           .           .           .           .           .
#> [16,]  -1.5858172   2.7650594  -0.4057733   0.3827061  -1.0918987   0.3059164
#> [17,]   0.1800004   0.7685879   0.9337187  -0.8282469  -1.0443035  -0.3438262
#> [18,]  -0.1452845  -0.2526594   0.6834694  -1.1656752   0.1582671   0.2830475
#> [19,]   0.5448889   1.7364790  -0.2394116  -0.7365064   0.4802055  -1.5056637
#> [20,]  -0.2674527   0.9884893   0.9025374   0.1915456  -1.5635116   0.2407236
metadata(se)
#> $title
#> [1] "Sample '.loom' file"
#>