Finding the nearest range neighbor in RangedSummarizedExperiment objects

This man page documents the nearest methods and family (i.e. precede, follow, distance, and distanceToNearest methods) for RangedSummarizedExperiment objects.

# S4 method for RangedSummarizedExperiment,ANY
precede(x, subject, select=c("arbitrary", "all"),
        ignore.strand=FALSE)
# S4 method for ANY,RangedSummarizedExperiment
precede(x, subject, select=c("arbitrary", "all"),
        ignore.strand=FALSE)

# S4 method for RangedSummarizedExperiment,ANY
follow(x, subject, select=c("arbitrary", "all"),
        ignore.strand=FALSE)
# S4 method for ANY,RangedSummarizedExperiment
follow(x, subject, select=c("arbitrary", "all"),
        ignore.strand=FALSE)

# S4 method for RangedSummarizedExperiment,ANY
nearest(x, subject, select=c("arbitrary", "all"), ignore.strand=FALSE)
# S4 method for ANY,RangedSummarizedExperiment
nearest(x, subject, select=c("arbitrary", "all"), ignore.strand=FALSE)

# S4 method for RangedSummarizedExperiment,ANY
distance(x, y, ignore.strand=FALSE, ...)
# S4 method for ANY,RangedSummarizedExperiment
distance(x, y, ignore.strand=FALSE, ...)

# S4 method for RangedSummarizedExperiment,ANY
distanceToNearest(x, subject, ignore.strand=FALSE, ...)
# S4 method for ANY,RangedSummarizedExperiment
distanceToNearest(x, subject, ignore.strand=FALSE, ...)

Arguments

x, subject
One of these two arguments must be a RangedSummarizedExperiment object.

select, ignore.strand
See ?nearest in the GenomicRanges package.

y
For the distance methods, one of x or y must be a RangedSummarizedExperiment object.

...
Additional arguments for methods.

Details

These methods operate on the rowRanges component of the RangedSummarizedExperiment object, which can be a GenomicRanges or GRangesList object.

More precisely, if any of the above functions is passed a RangedSummarizedExperiment object thru the x, subject, and/or y argument, then it behaves as if rowRanges(x), rowRanges(subject), and/or rowRanges(y) had been passed instead.

See ?nearest in the GenomicRanges package for the details of how nearest and family operate on GenomicRanges and GRangesList objects.

Value

See ?nearest in the GenomicRanges package.

See also

Examples

nrows <- 20; ncols <- 6
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
rowRanges <- GRanges(rep(c("chr1", "chr2"), c(5, 15)),
                     IRanges(sample(1000L, 20), width=100),
                     strand=Rle(c("+", "-"), c(12, 8)))
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
rse0 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                             rowRanges=rowRanges, colData=colData)
rse1 <- shift(rse0, 100)

res <- nearest(rse0, rse1)
res
#>  [1]  4  2  3  4  2 10  9 11  9 11 11 10 14 14 14 20 17 14 20 20
stopifnot(identical(res, nearest(rowRanges(rse0), rowRanges(rse1))))
stopifnot(identical(res, nearest(rse0, rowRanges(rse1))))
stopifnot(identical(res, nearest(rowRanges(rse0), rse1)))

res <- nearest(rse0)  # missing subject
res
#>  [1]  4  5  4  1  2 12  9 10  7  8 10  6 18 18 18 20 14 15 20 19
stopifnot(identical(res, nearest(rowRanges(rse0))))

hits <- nearest(rse0, rse1, select="all")
hits
#> Hits object with 40 hits and 0 metadata columns:
#>        queryHits subjectHits
#>        <integer>   <integer>
#>    [1]         1           4
#>    [2]         1           1
#>    [3]         2           2
#>    [4]         3           3
#>    [5]         4           4
#>    ...       ...         ...
#>   [36]        18          14
#>   [37]        18          18
#>   [38]        19          20
#>   [39]        19          19
#>   [40]        20          20
#>   -------
#>   queryLength: 20 / subjectLength: 20
stopifnot(identical(
  hits,
  nearest(rowRanges(rse0), rowRanges(rse1), select="all")
))
stopifnot(identical(
  hits,
  nearest(rse0, rowRanges(rse1), select="all")
))
stopifnot(identical(
  hits,
  nearest(rowRanges(rse0), rse1, select="all")
))